declared self.num_ghosts as int

+++ src/pyclaw/sharpclaw/solver.py
@@ -207,10 +207,10 @@ class SharpClawSolver(Solver):
         Allocate RK stage arrays or previous step solutions and fortran routine work arrays.
         """
         if self.lim_type == 2:
-            self.num_ghost = (self.reconstruction_order+1)/2
+            self.num_ghost = int((self.reconstruction_order+1)/2)
 
         if self.lim_type == 3:
-            self.num_ghost = 1 + self.reconstruction_order/2
+            self.num_ghost = int(1 + self.reconstruction_order/2)
 
         if self.lim_type == 2 and self.reconstruction_order != 5 and self.kernel_language == 'Python':
             raise Exception('Only 5th-order WENO reconstruction is implemented in Python kernels. \
